+++ src/Recyclarr.Cli/Pipelines/CustomFormat/Models/ConflictingCustomFormat.cs
@@ -1,5 +1,0 @@
-using Recyclarr.ResourceProviders.Domain;
-
-namespace Recyclarr.Cli.Pipelines.CustomFormat.Models;
-
-internal record ConflictingCustomFormat(CustomFormatResource GuideCf, int ConflictingId);