+++ src/Recyclarr.Cli/Pipelines/QualityProfile/Models/ConflictingQualityProfile.cs
@@ -1,5 +1,0 @@
-using Recyclarr.Cli.Pipelines.Plan;
-
-namespace Recyclarr.Cli.Pipelines.QualityProfile.Models;
-
-internal record ConflictingQualityProfile(PlannedQualityProfile PlannedProfile, int ConflictingId);