+++ src/Recyclarr.Cli/Processors/StateRepair/StateRepairResult.cs
@@ -1,9 +1,0 @@
-using Recyclarr.SyncState;
-
-namespace Recyclarr.Cli.Processors.StateRepair;
-
-internal record StateRepairResult(
-    List<TrashIdMapping> Mappings,
-    StateRepairStats Stats,
-    List<StateRepairDetail> Details
-);